My implementation of CS229
In this repo you will find my participation to CS229 mostly as a proof of work but also as I way to get familiar with GitHub.

Note that the best way to get as much as possible from this class is to do it the hard way, without looking at answers while completing each task. Nonetheless, after submission it can be a good practice to check for alternative implementations that can be more (i.e. vectorization vs loops).

Best of luck 😊